pyproject.toml
[project]
name = "clip-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "numpy>=2.2.4",
    "pillow>=11.1.0",
    "plotly>=6.0.1",
    "scikit-learn>=1.6.1",
]

[dependency-groups]
dev = ["jupyterlab>=4.4.0", "pre-commit>=4.2.0", "ruff>=0.11.4"]

[tool.ruff]
extend-exclude = ["./notebooks/TakeHomeTask.ipynb"]


[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    # pycodestyle
    "E",
    # pycodestyle warnings
    "W",
    # pyflakes
    "F",
    # bugbear
    "B",
    # comprehensions
    "C4",
    # implicit string concat
    "ISC",
    # tidy imports
    "TID",
    # isort
    "I",
    # pyupgrade
    "UP",
]


[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
